implementation slide view (helped by https://github.com/melongl11) and MVC Pattern appled in code and Overall Modified
