+++ app/src/main/AndroidManifest.xml
@@ -24,12 +24,15 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".MainActivity" />
         <activity android:name=".loginActivity.Create_user_activity" />
         <activity android:name=".getLatLng" />
         <activity android:name=".DBcheck" />
-        <activity android:name=".personalizingInfo" />
-        <activity android:name=".FeedbackActivity"></activity>
+        <activity android:name=".PreferenceController" />
+        <activity
+            android:name=".MainViewPager"
+            android:theme="@style/AppTheme">
+
+        </activity>
     </application>
 
 </manifest>
