+++ app/src/main/java/org/csw/narsi2/User.java
@@ -10,12 +10,15 @@ public class User implements Serializable {
     private String userID;
 
     private Feedback feedback;
+    private Preference preference;
 
     public User() {
-        feedback = new Feedback();
-
     }
 
+    public User(Feedback feedback, Preference preference) {
+        this.feedback = feedback;
+        this.preference = preference;
+    }
 
     public int getAge() {
         return age;
@@ -41,9 +44,23 @@ public class User implements Serializable {
         this.userID = userID;
     }
 
+    public void setFeedback(Feedback feedback) {
+        this.feedback = feedback;
+    }
+
+    public Preference getPreference() {
+        return preference;
+    }
+
+    public void setPreference(Preference preference) {
+        this.preference = preference;
+    }
+
     public Feedback getFeedback() {
         return feedback;
 
+
     }
+
 
 }
