--- app/src/main/res/layout/activity_main.xml
@@ -1,189 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<android.support.constraint.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:orientation="vertical"
-    tools:context=".MainActivity"
-    tools:layout_editor_absoluteY="25dp">
-
-
-    <ImageView
-        android:id="@+id/imageView"
-        android:layout_width="150dp"
-        android:layout_height="0dp"
-        android:layout_gravity="center"
-        android:layout_marginTop="14dp"
-        android:layout_marginEnd="11dp"
-        android:layout_marginBottom="23dp"
-        android:layout_weight="5"
-        android:background="#ffffff"
-        app:layout_constraintBottom_toTopOf="@+id/personalizing"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        app:srcCompat="@drawable/main_2" />
-
-    <ImageView
-        android:id="@+id/imageView3"
-        android:layout_width="150dp"
-        android:layout_height="0dp"
-        android:layout_gravity="center"
-        android:layout_marginStart="11dp"
-        android:layout_marginTop="2dp"
-        android:layout_marginBottom="2dp"
-        android:layout_weight="5"
-        android:background="#ffffff"
-        app:layout_constraintBottom_toBottomOf="@+id/imageView"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="@+id/imageView"
-        app:srcCompat="@drawable/title" />
-
-
-    <TextView
-        android:id="@+id/city"
-        android:layout_width="wrap_content"
-        android:layout_height="35dp"
-        android:layout_marginStart="37dp"
-        android:fontFamily="casual"
-        android:textSize="20dp"
-        app:layout_constraintBaseline_toBaselineOf="@+id/personalizing"
-        app:layout_constraintStart_toStartOf="parent" />
-
-    <TextView
-        android:id="@+id/personalizing"
-        android:layout_width="wrap_content"
-        android:layout_height="35dp"
-        android:layout_marginStart="9dp"
-        android:layout_marginBottom="28dp"
-        android:fontFamily="casual"
-        android:textSize="20dp"
-        app:layout_constraintBottom_toTopOf="@+id/wetRatio"
-        app:layout_constraintStart_toEndOf="@+id/city" />
-
-    <TextView
-        android:id="@+id/temp"
-        android:layout_width="wrap_content"
-        android:layout_height="35dp"
-        android:layout_marginStart="23dp"
-        android:fontFamily="casual"
-        android:textSize="20dp"
-        app:layout_constraintBaseline_toBaselineOf="@+id/personalizing"
-        app:layout_constraintStart_toEndOf="@+id/personalizing" />
-
-    <TextView
-        android:id="@+id/wetRatio"
-        android:layout_width="wrap_content"
-        android:layout_height="0dp"
-        android:layout_marginStart="8dp"
-        android:layout_marginEnd="8dp"
-        android:fontFamily="casual"
-        android:textSize="20dp"
-        app:layout_constraintBottom_toBottomOf="@+id/airPollution"
-        app:layout_constraintEnd_toStartOf="@+id/airPollution"
-        app:layout_constraintHorizontal_bias="0.008"
-        app:layout_constraintStart_toStartOf="@+id/city"
-        app:layout_constraintTop_toTopOf="@+id/airPollution"
-        app:layout_constraintVertical_bias="1.0" />
-
-    <TextView
-        android:id="@+id/airPollution"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="160dp"
-        android:layout_marginEnd="36dp"
-        android:fontFamily="casual"
-        android:textSize="20dp"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
-
-    <Button
-        android:id="@+id/logOut"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="16dp"
-        android:layout_marginBottom="57dp"
-        android:background="#ffffff"
-        android:text="로그아웃"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintStart_toStartOf="parent" />
-
-    <Button
-        android:id="@+id/dbcheck"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="8dp"
-        android:layout_marginTop="8dp"
-        android:layout_marginEnd="8dp"
-        android:layout_marginBottom="8dp"
-        android:background="#ffffff"
-        android:text="데이터베이스체크"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.846"
-        app:layout_constraintStart_toEndOf="@+id/logOut"
-        app:layout_constraintTop_toBottomOf="@+id/airPollution"
-        app:layout_constraintVertical_bias="0.84" />
-
-    <Button
-        android:id="@+id/pagechangebutton"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="8dp"
-        android:layout_marginTop="8dp"
-        android:layout_marginEnd="8dp"
-        android:layout_marginBottom="8dp"
-        android:background="#ffffff"
-        android:text="페이지이동"
-        app:layout_constraintBottom_toTopOf="@+id/logOut"
-        app:layout_constraintEnd_toStartOf="@+id/dbcheck"
-        app:layout_constraintHorizontal_bias="0.055"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/wetRatio"
-        app:layout_constraintVertical_bias="0.851" />
-
-    <Button
-        android:id="@+id/gojisu"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="8dp"
-        android:layout_marginTop="8dp"
-        android:layout_marginEnd="8dp"
-        android:layout_marginBottom="8dp"
-        android:text="Button"
-        app:layout_constraintBottom_toTopOf="@+id/dbcheck"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.772"
-        app:layout_constraintStart_toEndOf="@+id/pagechangebutton"
-        app:layout_constraintTop_toBottomOf="@+id/airPollution"
-        app:layout_constraintVertical_bias="0.851" />
-
-    <android.support.constraint.ConstraintLayout
-        android:id="@+id/layout_waiting"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:background="#BB000000"
-        android:elevation="3dp"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent">
-
-        <ProgressBar
-            android:id="@+id/progressBar"
-            style="?android:attr/progressBarStyle"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginStart="8dp"
-            android:layout_marginTop="8dp"
-            android:layout_marginEnd="8dp"
-            android:layout_marginBottom="8dp"
-            app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toTopOf="parent" />
-    </android.support.constraint.ConstraintLayout>
-
-
-</android.support.constraint.ConstraintLayout>
-
