+++ local.properties
@@ -4,5 +4,5 @@
 # Location of the SDK. This is only used by Gradle.
 # For customization when using a Version Control System, please read the
 # header note.
-#Mon Oct 29 20:29:21 KST 2018
+#Wed Dec 05 17:32:22 KST 2018
 sdk.dir=C\:\\Users\\alsrn\\AppData\\Local\\Android\\Sdk
